extract all sections of stable.js to separate module
